requirements.txt
certifi==2024.2.2
chardet==3.0.4
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy==1.26.4
pillow==10.2.0
rfc3986==1.5.0
sniffio==1.3.1
tkVideo==0.1
